+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/utils/strategies/BaseStrategy.js
@@ -1,4 +1,7 @@
+const { Logger } = require('@cerner/terra-cli');
 const CropDimension = require('../CropDimension');
+
+const logger = new Logger({ prefix: '[wdio-visual-regression-service:BaseStrategy]' });
 
 class BaseStrategy {
   constructor(screenDimensions) {
@@ -13,25 +16,36 @@ class BaseStrategy {
   }
 
   setScrollArea(startX, startY, endX, endY) {
-    // TODO: bug https://github.com/zinserjan/wdio-screenshot/issues/87
-    // const documentWidth = this.screenDimensions.getDocumentWidth();
-    // const documentHeight = this.screenDimensions.getDocumentHeight();
+    const documentWidth = this.screenDimensions.getDocumentWidth();
+    const documentHeight = this.screenDimensions.getDocumentHeight();
 
-    // if (startX >= documentWidth) {
-    //   throw new Error('startX is out of range');
-    // } else if (startY >= documentHeight) {
-    //   throw new Error('startY is out of range');
-    // } else if (endX > documentWidth) {
-    //   throw new Error('endX is out of range');
-    // } else if (endY > documentHeight) {
-    //   throw new Error('endY is out of range');
-    // }
+    // Capture what is visible if the element is rendered partially off the screen.
+    // Capture the entire viewable screen dimension if it is rendered completely off the screen.
+    const adjustedStartX = startX > documentWidth || startX < 0 ? 0 : startX;
+    const adjustedStartY = startY > documentHeight || startY < 0 ? 0 : startY;
+    const adjustedEndX = endX > documentWidth || endX < 0 ? documentWidth : endX;
+    const adjustedEndY = endY > documentHeight || endY < 0 ? documentHeight : endY;
+
+    if (adjustedStartX !== startX || adjustedStartY !== startY || adjustedEndX !== endX || adjustedEndY !== endY) {
+      const dimensions = {
+        startX,
+        startY,
+        endX,
+        endY,
+        adjustedStartX,
+        adjustedStartY,
+        adjustedEndX,
+        adjustedEndY,
+      };
+
+      logger.info(`The element being captured is off the viewable screen and will be clipped using the adjusted dimensions: ${JSON.stringify(dimensions, null, 2)}`);
+    }
 
     this.area = {
-      startX,
-      startY,
-      endX,
-      endY,
+      startX: adjustedStartX,
+      startY: adjustedStartY,
+      endX: adjustedEndX,
+      endY: adjustedEndY,
     };
   }